Change dyt to element-wise op
